03-agent-with-different-models-and-tools-crewai/requirements.txt
crewai>=0.28.0
langchain>=0.1.0
langchain-google-genai>=0.0.7
langchain-openai>=0.0.5
langchain-anthropic>=0.0.2
crewai-tools>=0.1.2
python-dotenv>=1.0.0
ipykernel>=6.0.0
jupyter>=1.0.0
